+++ CHANGELOG.md
@@ -1,9 +1,20 @@
+## 0.1.2
+
+Github release:
+
+- Support includePrerelease, default is false.
+
+Commands:
+
+- Update `choice` to `choose`.
+- Add alias `c` for `choose`.
+
 ## 0.1.1
 
-- support overwrite for directory
+- Support overwrite for directory.
 
 ## 0.1.0
 
 The first version
 
-- support github release
+- Support github release